+++ src/scripts/routes/routes.js
@@ -1,5 +1,5 @@
-import About from '../../views/page/about';
-import Home from '../../views/page/home';
+import About from '../../views/page/about_2';
+import Home from '../../views/page/home_2';
 
 const routes = {
   '/': Home,
