+++ src/views/page/home_2.js
